split RequestController to extract vscode dependencies
